Add files via upload
Portobello Tech is an app innovator who has devised an intelligent way of 
predicting employee turnover within the company. It periodically evaluates 
employees' work details, including the number of projects they worked on, 
average monthly working hours, time spent in the company, promotions in the 
last five years, and salary level. 
Data from prior evaluations shows the employees’ satisfaction in the workplace. 
The data could be used to identify patterns in work style and their interest in 
continuing to work for the company.  
The HR Department owns the data and uses it to predict employee turnover. 
Employee turnover refers to the total number of workers who leave a company 
over time. 
As the ML Developer assigned to the HR Department, you have been asked to 
create ML programs